+++ api/app/hfi/hfi_calc.py
@@ -9,7 +9,9 @@ from sqlalchemy.orm import Session
 import app
 from app.db.database import get_read_session_scope
 from app.db.models.hfi_calc import PlanningWeatherStation, FuelType as FuelTypeModel
-from app.fire_behaviour.prediction import calculate_fire_behaviour_prediction, FireBehaviourPrediction
+from app.fire_behaviour.cffdrs import CFFDRSException
+from app.fire_behaviour.prediction import (
+    FireBehaviourPredictionInputError, calculate_fire_behaviour_prediction, FireBehaviourPrediction)
 from app.schemas.hfi_calc import (DailyResult, DateRange,
                                   FireStartRange, HFIResultRequest,
                                   PlanningAreaResult,
@@ -63,14 +65,8 @@ def generate_station_daily(raw_daily: dict,  # pylint: disable=too-many-locals
             pdf=pdf,
             cbh=cbh,
             cfl=cfl)
-    # pylint: disable=broad-except
-    except Exception as exc:
-        # TODO: Remove this exception - it can hide away bugs in code. Catch more specific exceptions.
-        #   e.g.: for c7b, if cc is null, we can't calculate - so let's throw a specific exception and
-        #   catch that.
-        logger.error('Encountered error while generating StationDaily for station %s', station.code)
-        logger.error(exc, exc_info=True)
-        # prediction calculation failed, so we set the values to None
+    except (FireBehaviourPredictionInputError, CFFDRSException) as error:
+        logger.info("Error calculating fire behaviour prediction for station %s : %s", station.code, error)
         fire_behaviour_prediction = FireBehaviourPrediction(None, None, None, None, None)
 
     return StationDaily(
@@ -177,7 +173,6 @@ async def hydrate_fire_centres():
             # Combine everything.
             station_properties = WeatherStationProperties(
                 name=wfwx_station.name,
-                fuel_type=station_info['fuel_type'],
                 elevation=wfwx_station.elevation,
                 wfwx_station_uuid=wfwx_station.wfwx_station_uuid)