+++ web/src/features/hfiCalculator/util.ts
@@ -1,7 +1,8 @@
-import { StationDaily, PlanningArea } from 'api/hfiCalculatorAPI'
+import { StationDaily, PlanningArea, FuelType } from 'api/hfiCalculatorAPI'
 import {
   HFIResultResponse,
-  PrepDateRange
+  PrepDateRange,
+  StationInfo
 } from 'features/hfiCalculator/slices/hfiCalculatorSlice'
 import { groupBy, isUndefined, sortBy, take } from 'lodash'
 import { DateTime } from 'luxon'
@@ -63,19 +64,47 @@ export const getDailiesByStationCode = (
   return dailiesForCode ? dailiesForCode : []
 }
 
+export const getPlanningAreaStationInfo = (
+  planning_area_station_info: { [key: number]: StationInfo[] } | undefined,
+  planningAreaId: number,
+  code: number
+): StationInfo | undefined => {
+  if (
+    !isUndefined(planning_area_station_info) &&
+    !isUndefined(planning_area_station_info[planningAreaId])
+  ) {
+    return planning_area_station_info[planningAreaId].find(
+      info => info.station_code === code
+    )
+  }
+  return undefined
+}
+
 export const stationCodeSelected = (
-  result: HFIResultResponse | undefined,
+  planning_area_station_info: { [key: number]: StationInfo[] },
   planningAreaId: number,
   code: number
 ): boolean => {
-  if (
-    !isUndefined(result) &&
-    !isUndefined(result.planning_area_station_info[planningAreaId])
-  ) {
-    const station_info = result.planning_area_station_info[planningAreaId].find(
-      info => info.station_code === code
-    )
-    return station_info?.selected ?? false
-  }
-  return false
+  const stationInfo = getPlanningAreaStationInfo(
+    planning_area_station_info,
+    planningAreaId,
+    code
+  )
+  return stationInfo?.selected ?? false
+}
+
+export const getSelectedFuelType = (
+  planningAreaStationInfo: { [key: number]: StationInfo[] } | undefined,
+  planningAreaId: number,
+  stationCode: number,
+  fuelTypes: FuelType[]
+) => {
+  const stationInfo = getPlanningAreaStationInfo(
+    planningAreaStationInfo,
+    planningAreaId,
+    stationCode
+  )
+  return isUndefined(stationInfo)
+    ? undefined
+    : fuelTypes.find(instance => instance.id == stationInfo.fuel_type_id)
 }